Move WebAssemblyRsaExtensions to a separate file
