+++ Source/Blazorise/Licensing/BlazoriseLicenseProvider.cs
@@ -351,13 +351,3 @@ public sealed class BlazoriseLicenseProvider
 
     #endregion
 }
-
-internal static class WebAssemblyRsaExtensions
-{
-    public static IVerifier_LoadAndVerify WithWebAssemblyRsaPublicKey( this IVerifier_WithVerifier signer, IJSRuntime jsRuntime, IVersionProvider versionProvider, string base64EncodedCsbBlobKey )
-    {
-        var rsaVerifier = new WebAssemblyRsaVerifier( jsRuntime, versionProvider, base64EncodedCsbBlobKey );
-        signer.WithVerifier( rsaVerifier );
-        return signer as IVerifier_LoadAndVerify;
-    }
-}
